refactor: remove the provisioning-hook feature
No app ships a provision.sh any more. halos-marine-containers#210 moved the
Signal K plugin set into a pre-baked image, which was the feature's only
consumer, so the generator, its one-shot unit template, the packaging rules and
the whole hook contract have nothing left to serve. Restoring it from git is
cheaper than carrying it unused.

Breaking for the generator's supported app shape: an app that still ships a
provision.sh now gets it silently ignored rather than packaged. Nothing in the
fleet does.

What survives, deliberately:

- StartLimitIntervalSec/StartLimitBurst in service.j2 are from #216 and are
  unrelated to provisioning. They get *stronger* here -- the DESIGN.md note
  saying the burst stops being a reliable "this is broken" signal existed
  because waiting on the one-shot inflated each start attempt, so it goes.
- The file-watcher blocks sit directly adjacent to the removed ones in rules.j2
  and prerm.j2. Structurally independent, untouched.
- The unconditional --no-start / restart --no-block behaviour is not
  provisioning-specific despite comments that named it. The comments are
  reworded to describe a slow ExecStartPre instead, and the two affected golden
  fixtures regenerated in the same commit -- the diff is comment-only.
- TestInstallTimeStart and the hookless golden suite were the only tests in
  test_provisioning.py not about provisioning. Rehomed to
  tests/test_generated_output.py, with the PROVISION_APP fixture references
  converted to the simple app.

The golden suite is what makes this safe to land: removing a conditional block
can shift surrounding whitespace for every app while substring assertions stay
green. It passed byte-identical before the comment rewording.

Closes #240. Closes #239 -- overlapping restarts aborting an in-flight
provisioning run is unreachable once no app ships a hook.

918 passed, 18 skipped.
