+++ EXAMPLES.md
@@ -25,7 +25,6 @@ my-app/
 ├── docker-compose.yml     # Required: Container orchestration
 ├── config.yml             # Required: User configuration schema
 ├── prestart.sh            # Optional: app-specific startup hook (see below)
-├── provision.sh           # Optional: one-time setup hook, own unit (see below)
 ├── assets/                # Optional: static payload installed to <lib>/assets/
 ├── default-data/          # Optional: static seed files (baked at build time)
 ├── icon.png               # Optional: Application icon (PNG or SVG)
@@ -38,20 +37,10 @@ A `prestart.sh` is **not** a replacement for the generated prestart — the gene
 app-specific logic in it (secrets, config seeding), append to `$RUNTIME_ENV` with `>>`, and prefer
 `default-data/` for static seed files. See the Custom Prestart Hook contract in `AGENTS.md`.
 
-A `provision.sh` is for setup the app cannot run without — installing plugins, fetching data. It gets
-its own `<package>-provision.service` which the app `Requires=`, so **the app does not start until the
-hook succeeds**: a half-provisioned application is worse than one that is plainly down. Nothing
-outside the hook retries it, so the hook retries internally and exits non-zero only when waiting
-cannot help. Two differences from `prestart.sh` matter: it is **executed, not sourced** (its own
-process, before the prestart runs), and it runs on **every** app start, so it must be idempotent and
-a fast no-op when there is nothing to do. Anything genuinely optional does not belong here. See the
-Provisioning Hook contract in `AGENTS.md`.
-
 Static payload a hook reads at runtime (a package manifest, a template) belongs in `assets/`, which
-is installed to `/var/lib/container-apps/<package>/assets/`. A provisioning hook can reach it as
-plain `assets/<file>`, since its unit sets `WorkingDirectory` to the package lib dir. Do not leave
-such a file at the top level of the app directory: only the files listed above are packaged, so a
-stray one silently never ships.
+is installed to `/var/lib/container-apps/<package>/assets/`. A prestart hook can reach it as
+`/var/lib/container-apps/<package>/assets/<file>`. Do not leave such a file at the top level of the
+app directory: only the files listed above are packaged, so a stray one silently never ships.
 
 ### Basic Workflow
