+++ docs/DESIGN.md
@@ -280,7 +280,6 @@ output-directory/
 │   │   ├── copyright            # Copyright and license
 │   │   ├── compat               # Debhelper compatibility level
 │   │   ├── <package>.service    # systemd service file
-│   │   └── <package>-provision.service   # Only when the app ships provision.sh
 │   ├── docker-compose.yml       # Copied from input
 │   ├── config.yml               # Copied from input
 │   ├── metadata.yaml            # Copied from input
@@ -523,9 +522,7 @@ WantedBy=multi-user.target
 
 **Note on StartLimit and recovery**: `Restart=always` alone makes a deterministically-failing `ExecStartPre` (the prestart, `configure-container-routing`, or an `app-prestart.sh` hook running under `set -e`) retry every `RestartSec` forever. `StartLimitIntervalSec=600` / `StartLimitBurst=5` bound this: 5 failed starts within 600s drive the unit to `failed` (~40s with `RestartSec=10`) instead of looping silently, while legitimate in-app restarts spaced further apart than `interval/burst` never accumulate. A unit that hits the start limit stays in `failed` and does **not** self-heal — recover with `systemctl reset-failed <package>.service && systemctl start <package>.service` after fixing the underlying cause. A package upgrade does clear it in practice: postinst runs `systemctl restart --no-block <package>.service`, which starts an inactive unit on first install and restarts a running one on upgrade.
 
-**Note on the install-time start**: postinst issues that restart itself, and `rules` passes `--no-start` to `dh_installsystemd` to suppress debhelper's own. debhelper's snippet calls `deb-systemd-invoke`, which runs `systemctl start` without `--no-block` and therefore waits for the unit's whole start transaction while dpkg holds its lock. For an app with a provisioning one-shot ordered ahead of it, that is the full provisioning run — minutes, on a cold cache. `--no-block` returns as soon as the job is queued; ordering is a property of the units, so the app still starts after provisioning completes.
-
-**Note on StartLimit for apps with a provisioning hook**: waiting on `<package>-provision.service` inflates each start attempt from ~`RestartSec` to provision-duration + `RestartSec`. Once provisioning routinely exceeds ~120s, five attempts no longer fit inside `StartLimitIntervalSec=600`, the burst stops tripping, and a deterministically-failing `ExecStartPre` restarts indefinitely instead of terminating in `failed`. This is accepted rather than fixed: scaling the interval to match the provisioning backstop would make legitimate spaced-out restarts accumulate toward the burst. For provisioned apps, treat a repeating restart loop — not a `failed` unit — as the signal that a prestart is broken.
+**Note on the install-time start**: postinst issues that restart itself, and `rules` passes `--no-start` to `dh_installsystemd` to suppress debhelper's own. debhelper's snippet calls `deb-systemd-invoke`, which runs `systemctl start` without `--no-block` and therefore waits for the unit's whole start transaction while dpkg holds its lock — for an app with a slow `ExecStartPre`, that stalls the package transaction behind it. `--no-block` returns as soon as the job is queued.
 
 ## Path Conventions
