+++ src/generate_container_packages/templates/debian/rules.j2
@@ -25,11 +25,6 @@ override_dh_auto_install:
 	install -D -m 755 app-prestart.sh \
 		debian/{{ package.name }}/{{ paths.lib }}/app-prestart.sh
 {% endif %}
-{% if has_provision %}
-	# Install the provisioning hook (executed by {{ package.name }}-provision.service)
-	install -D -m 755 provision.sh \
-		debian/{{ package.name }}/{{ paths.lib }}/provision.sh
-{% endif %}
 
 {% if has_icon %}
 	# Install icon
@@ -83,12 +78,6 @@ override_dh_auto_install:
 	# Install systemd service file
 	install -D -m 644 debian/{{ package.name }}.service \
 		debian/{{ package.name }}/{{ paths.systemd }}/{{ package.name }}.service
-{% if has_provision %}
-
-	# Install the provisioning unit (pulled in by the app unit's Requires=, so not enabled)
-	install -D -m 644 debian/{{ package.name }}-provision.service \
-		debian/{{ package.name }}/{{ paths.systemd }}/{{ package.name }}-provision.service
-{% endif %}
 
 {% if has_file_watchers %}
 	# Install file watcher systemd units
@@ -110,7 +99,7 @@ override_dh_auto_test:
 override_dh_installsystemd:
 	# debhelper's own start call blocks dpkg: deb-systemd-invoke runs
 	# `systemctl start` without --no-block, so it waits for the unit's whole
-	# start transaction. For an app with a provisioning one-shot ordered ahead
-	# of it that can be many minutes, all while dpkg holds its lock. postinst
-	# starts the service itself with --no-block instead.
+	# start transaction. For an app with a slow ExecStartPre that can be many
+	# minutes, all while dpkg holds its lock. postinst starts the service
+	# itself with --no-block instead.
 	dh_installsystemd --no-start
