--- src/generate_container_packages/templates/systemd/provision-service.j2
@@ -1,47 +1,0 @@
-[Unit]
-Description=Provisioning for {{ package.human_name }}
-# Separate unit so neither provisioning's runtime nor its failures count against
-# {{ package.name }}.service's start limit.
-After=docker.service
-Requires=docker.service
-# docker.service being active says nothing about DNS or a default route, so
-# without this a device with working internet can lose the boot race and be
-# indistinguishable from an offline one.
-Wants=network-online.target
-After=network-online.target
-# Stopping the app stops an in-flight run; without this the hook keeps writing to
-# the app's data directory after the operator sees the app as stopped.
-PartOf={{ package.name }}.service
-# The app pulls this unit in on every one of its start attempts, so a crash-looping
-# app would otherwise exhaust the default start limit here and have its starts
-# refused -- silently, since nothing reports a refused dependency job.
-StartLimitIntervalSec=0
-
-[Service]
-# Deliberately no Restart=. A restart job against this unit propagates through the
-# app's Requires= and stops the running app, so retrying here would tear down the
-# very thing provisioning exists to bring up. The hook retries internally instead
-# and exits only when it is finished or has given up. No RemainAfterExit= either:
-# returning to inactive is what lets the hook re-run on each app start.
-Type=oneshot
-WorkingDirectory={{ service.working_directory }}
-# The app's prestart has not run yet, so runtime.env does not exist.
-EnvironmentFile=-{{ service.env_defaults_file }}
-EnvironmentFile=-{{ service.env_file }}
-# The hook must name any container it runs "$HALOS_PROVISION_CONTAINER" so the
-# cleanup below can reap it. Stopping the app signals this unit's cgroup, so an
-# untrapped hook dies mid-run without its own cleanup, while a container started
-# by dockerd outlives the unit and keeps writing to the data volume. ExecStopPost
-# still runs on that path, and reads the same variable so an EnvironmentFile
-# override cannot make the reaper and the hook disagree about the name.
-Environment=HALOS_PROVISION_CONTAINER={{ package.name }}-provision
-ExecStart={{ service.working_directory }}/provision.sh
-ExecStopPost=-docker rm -f ${HALOS_PROVISION_CONTAINER}
-# No outer bound: the app is gated on this unit, so a timeout would kill the hook
-# and leave the app permanently down with nothing left to retry. Waiting is the
-# lesser failure. The hook owns every bound -- per-operation timeouts, its retry
-# schedule, and when to give up -- and a hook that gives up exits non-zero, which
-# surfaces as a failed unit rather than an indefinite wait.
-TimeoutStartSec=infinity
-StandardOutput=journal
-StandardError=journal
