+++ src/generate_container_packages/templates/systemd/service.j2
@@ -19,23 +19,6 @@ Wants=halos-authelia-container.service
 After=halos-resolve-domain.service
 Wants=halos-resolve-domain.service
 {% endif %}
-{% if has_provision %}
-# Requires=, not Wants=: an app that ships a provisioning hook needs what the hook
-# installs, so starting without it would leave a randomly degraded installation
-# rather than a working one. A failed run therefore withholds the app, and the
-# provision unit retries until it succeeds and then starts this unit itself --
-# a dependency failure leaves no start job here for Restart= to act on.
-Requires={{ package.name }}-provision.service
-# Accepted consequence: waiting on that one-shot inflates each start attempt from
-# ~RestartSec to provision-duration + RestartSec. Once provisioning routinely
-# exceeds ~120s, five attempts no longer fit in StartLimitIntervalSec above, the
-# burst stops tripping, and a deterministically-failing ExecStartPre restarts
-# indefinitely instead of terminating in `failed`. Scaling the interval to match
-# would make legitimate spaced-out restarts accumulate toward the burst, so the
-# limit is left as-is for provisioned apps and the weaker guarantee is documented
-# in docs/DESIGN.md.
-After={{ package.name }}-provision.service
-{% endif %}
 
 [Service]
 Type=simple
