+++ tests/fixtures/expected/simple-app.rules
@@ -52,7 +52,7 @@ override_dh_auto_test:
 override_dh_installsystemd:
 	# debhelper's own start call blocks dpkg: deb-systemd-invoke runs
 	# `systemctl start` without --no-block, so it waits for the unit's whole
-	# start transaction. For an app with a provisioning one-shot ordered ahead
-	# of it that can be many minutes, all while dpkg holds its lock. postinst
-	# starts the service itself with --no-block instead.
+	# start transaction. For an app with a slow ExecStartPre that can be many
+	# minutes, all while dpkg holds its lock. postinst starts the service
+	# itself with --no-block instead.
 	dh_installsystemd --no-start
