--- tests/fixtures/valid/provision-app/assets/seed.list
@@ -1,4 +1,0 @@
-# Payload the provisioning hook reads at runtime.
-# Present so packaging tests can assert that assets/ and provision.sh
-# land in the same .deb.
-example-package
