--- tests/fixtures/valid/provision-app/config.yml
@@ -1,26 +1,0 @@
-version: "1.0"
-groups:
-  - id: general
-    label: General Settings
-    description: Basic application configuration
-    fields:
-      - id: APP_PORT
-        label: Application Port
-        type: integer
-        default: 8080
-        required: true
-        min: 1024
-        max: 65535
-        description: Port on which the application will listen
-
-      - id: LOG_LEVEL
-        label: Log Level
-        type: enum
-        default: info
-        required: false
-        options:
-          - debug
-          - info
-          - warning
-          - error
-        description: Logging verbosity level
