--- tests/fixtures/valid/provision-app/docker-compose.yml
@@ -1,17 +1,0 @@
-version: '3.8'
-
-services:
-  app:
-    image: nginx:alpine
-    container_name: provision-test-app
-    ports:
-      - "${APP_PORT:-8080}:80"
-    environment:
-      - LOG_LEVEL=${LOG_LEVEL:-info}
-    volumes:
-      - /var/lib/container-apps/provision-test-app-container/data:/usr/share/nginx/html:rw
-    restart: unless-stopped
-    logging:
-      driver: journald
-      options:
-        tag: "{{.Name}}"
