--- tests/fixtures/valid/provision-app/metadata.yaml
@@ -1,22 +1,0 @@
-# Provision App - exercises the provisioning hook (provision.sh)
-name: Provision Test App
-app_id: provision-test-app
-version: 1.0.0
-upstream_version: 1.0.0
-description: A test application with a provisioning hook
-long_description: |
-  This application ships a provision.sh hook, so the generator emits a
-  decoupled one-shot provisioning unit that the app service orders after.
-homepage: https://example.com/provision-app
-maintainer: Test Developer <test@example.com>
-license: MIT
-tags:
-  - role::container-app
-  - implemented-in::docker
-debian_section: net
-architecture: all
-
-# Default configuration
-default_config:
-  APP_PORT: "8080"
-  LOG_LEVEL: "info"
